Added a functionality for deleting your own photos
Some code cleanup was done as well. Now using `render` instead of
`render_to_response`. Some URLs were changed. Also had to undo the
last commit, because then the ImageField no longer has the width
and height attributes, which are needed for selecting a region of
interest. Also started using HttpResponseRedirect when form data
was successfully processed, as adviced in the Django documentation.