Add an exception for configurations that do not affect features
The option `grabcut.output_folder` is part of the `preprocess`
configuration, but this setting should not trigger caching of
new features, because it does not affect features extracted. This
option no longer triggers feature extraction.

And to prevent repeation of code, the FingerprintCache class was
moved to the nbclassify package.