Update nbc-trainer.
Added option to give a codebook to the classifier. 
A codebook is necessary if the BagOfWords algorithm was applied to the data the ann was trained with. The features of the image that needs to be classified first need to be compared to the codebook to create the correct code for the ann.