Added codebooklength check, removed logging info.
Sometimes a codebook is created with less clusters than wanted, possibly because of centroids without datapoints in the kmeans method when there are little datapoints to use. In that case the codebook must be created again until it has the right amount of clusters.

Also some logging information that was added for testing/debugging was removed.