Fixed some bugs.
If there was only one class, an error occurred because there was no [photo_id] entry in self.classifications. The code that creates this entry is now placed before this class check on line 580-582. 
If there was only on class, another error occurred because the ann was not created yet, but needed to be destroyed. The code to create the ann is now placed before the class check on line 552.