Additions and changes to aid compatibility with updated packages.
User will now be informed of images that do not meet the requirements to find all corners and the saved information will be deleted from the database. This ensures that a new photo can be made and the image(s) can be uploaded again with the same information.

Newest version of Django (1.11) is not compatible with latest version of Geoposition, changes are made to aid in compatibility with an unreleased new version of Geoposition.

Changes are made to guide Django to the right folders, which needed to be changed  due to the changes made in the newest version of Django (1.11)